Move `ChunkId`, `RowRange` and `RowIdx` out of the `layout` module
